+++ treeVectorDirection.h
@@ -16,7 +16,7 @@ class treeVectorDirection:public directionFunction {
 public:
     treeVectorDirection(dataManager* data,int nLeaves, int minimumNodeSize, _TREE_TYPE_ treeType);
     virtual ~treeVectorDirection();
-    void eval(double* pnt, double* direction);
+    void eval(float * pnt, float * direction);
     void buildDirection();
     void saveTree(ofstream& fileDB);
     //Data Manager
@@ -27,34 +27,34 @@ public:
 private:
     class NODE{
     public:
-        NODE(dataManager* data,treeVectorDirection* tree,int leftPoint,int rightPoint, double loss);
+        NODE(dataManager* data,treeVectorDirection* tree,int leftPoint,int rightPoint, float  loss);
         ~NODE();
         int _leftPoint;
         int _rightPoint;
         NODE* _leftChildNode;
         NODE* _rightChildNode;
         //loss
-        double _nodeLoss;
+        float  _nodeLoss;
         //gain of current node
-        double _nodeGain;
+        float  _nodeGain;
         //raw _nodeSumH   --> sum of Hessian elements
         //    _nodeSumG   --> sum of Gradient elements
-        double* _nodeSumH;
-        double* _nodeSumG;
+        float * _nodeSumH;
+        float * _nodeSumG;
         //best split gain with this node
-        double _additiveGain;
+        float  _additiveGain;
         //whether this point is an internal point
         bool _isInternal;
         int _iDimension;
-        double _cut;
+        float  _cut;
         //current working class
         int _class;
         //regression value of current node
-        double _f;
+        float  _f;
         bool _ableSplit;
         dataManager* _data;
         treeVectorDirection* _tree;
-        void bestNode(NODE*& n,double& gain);
+        void bestNode(NODE*& n,float & gain);
         void splitNode();
         void saveNode(ofstream& fileDB);
         bool printInfo(const char* indent,bool last);
@@ -63,12 +63,12 @@ private:
 
         //variables for node split
         int _nG;
-        double *leftSumG, *leftSumH;
-        double *rightSumG, *rightSumH;
-        double *leftSumG1, *leftSumH1;
-        double *rightSumG1, *rightSumH1;
-        double leftLoss, rightLoss;
-        double leftLoss1, rightLoss1;
+        float  *leftSumG, *leftSumH;
+        float  *rightSumG, *rightSumH;
+        float  *leftSumG1, *leftSumH1;
+        float  *rightSumG1, *rightSumH1;
+        float  leftLoss, rightLoss;
+        float  leftLoss1, rightLoss1;
         
     };
     
@@ -78,7 +78,7 @@ private:
     //tree array for different classes
     NODE* _rootNode;
     //tree evaluation
-    double evalp(double* s,int& iClass);
+    float  evalp(float * s,int& iClass);
     //initialize root nodes
     void initNode();
     void reArrange(NODE* node, int splitPoint);
